Totals for offer are now calculated and saved to db

+++ src/Sparrow.Web/Models/Drafts/DraftItemAddModel.cs
@@ -5,7 +5,7 @@ namespace Sparrow.Web.Models.Drafts
     public class DraftItemAddModel
     {
         public int Quantity { get; set; }
-        public double Discount { get; set; }
+        public int Discount { get; set; }
         public Guid? ProductId { get; set; }
     }
 }
